fix: set deletion timestamp in strategy
A previous commit changed the strategy such that it keys off of the
deletion timestamp and finalizers for setting the deleted field. This
broke the test because the test calls the strategy directly with no
deletion timestamp set.

With this change, the delete method on the db strategy will ensure that
the deletion timestamp is set.

Signed-off-by: Donnie Adams <donnie@acorn.io>